requirements.txt
psycopg2==2.9.4
psycopg2_binary==2.9.3
python-decouple==3.6
